Added settings and capabilities for 0.4.1
Added logic around the controls for ENABLE_CONDITIONS, ENABLE_TOKEN_IMAGES, and DAMAGE_ADDS. The first of these gates the application of conditions (AGE system specific), the second enables/disables the inclusion of token images in the dialog. The last is a big jump to try to make things compatible for games like SWADE and L5R5E where damage is additive from a base of 0 rather than subtracting from a pool of health.