Update token-health.js
Removed the logic that limited launching the dialog box based on the number of controlled tokens - that seems to break things such that once the dialog box doesn't launch for this reason it will never launch even when it should.

Added logic to restrict launching the dialog based on if the user is a Player and the GM has disabled use by players.

Commented out code that toggles the restricted setting for the keybinding.