Update token-health.js
Moved assembly of affected tokens array to the toggle function to allow culling to only those tokens owned by the user. This enables the module to not launch the dialog when no tokens would be affected . The toggle function now passes in the array of tokens to be affected to the displayOverlay function.

Added logic to produce a UI warning to players if all of the tokens they've targeted are not owned by them. If only some are owned, then no warning is generated, but the list is culled and Token Health will only show those tokens they own in the affected list.